# SklepInternetowy/SklepInternetowy/OpenAPIs/swagger3.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RestApiZamowienia",
    "version": "1.0"
  },
  "paths": {
    "/api/Adres": {
      "get": {
        "tags": [
          "Adres"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Adre"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Adre"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Adre"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Adres"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Adre"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Adre"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Adre"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Adre"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Adre"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Adre"
                }
              }
            }
          }
        }
      }
    },
    "/api/Adres/{id}": {
      "get": {
        "tags": [
          "Adres"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Adre"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Adre"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Adre"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Adres"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Adre"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Adre"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Adre"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Adres"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/MetodaPlatnoscis": {
      "get": {
        "tags": [
          "MetodaPlatnoscis"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetodaPlatnosci"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetodaPlatnosci"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetodaPlatnosci"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MetodaPlatnoscis"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetodaPlatnosci"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MetodaPlatnosci"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MetodaPlatnosci"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MetodaPlatnosci"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetodaPlatnosci"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetodaPlatnosci"
                }
              }
            }
          }
        }
      }
    },
    "/api/MetodaPlatnoscis/{id}": {
      "get": {
        "tags": [
          "MetodaPlatnoscis"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MetodaPlatnosci"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetodaPlatnosci"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetodaPlatnosci"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "MetodaPlatnoscis"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetodaPlatnosci"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MetodaPlatnosci"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MetodaPlatnosci"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "MetodaPlatnoscis"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Adre": {
        "type": "object",
        "properties": {
          "idAdresu": {
            "type": "integer",
            "format": "int32"
          },
          "kraj": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "miejscowosc": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "ulica": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "kodPocztowy": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "nrBudynku": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "nrLokalu": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "ktoDodal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kiedyDodano": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ktoEdytowal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kiedyEdytowano": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ktoUsunal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kiedyUsunieto": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "aktywny": {
            "type": "boolean",
            "nullable": true
          },
          "klients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Klient"
            },
            "nullable": true
          },
          "ktoDodalNavigation": {
            "$ref": "#/components/schemas/Uzytkownik"
          },
          "ktoEdytowalNavigation": {
            "$ref": "#/components/schemas/Uzytkownik"
          },
          "ktoUsunalNavigation": {
            "$ref": "#/components/schemas/Uzytkownik"
          }
        },
        "additionalProperties": false
      },
      "ElementKoszyka": {
        "type": "object",
        "properties": {
          "idElementuKoszyka": {
            "type": "integer",
            "format": "int32"
          },
          "idSesjiKoszyka": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idTowaru": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ilosc": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idSesjiKoszykaNavigation": {
            "$ref": "#/components/schemas/SesjaKoszyka"
          },
          "idTowaruNavigation": {
            "$ref": "#/components/schemas/Towar"
          }
        },
        "additionalProperties": false
      },
...
      "Kategorium": {
        "type": "object",
        "properties": {
          "idKategorii": {
            "type": "integer",
            "format": "int32"
          },
          "nazwa": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "ktoDodal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kiedyDodano": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ktoEdytowal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kiedyEdytowano": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ktoUsunal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kiedyUsunieto": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "aktywny": {
            "type": "boolean",
            "nullable": true
          },
          "ktoDodalNavigation": {
            "$ref": "#/components/schemas/Uzytkownik"
          },
          "ktoEdytowalNavigation": {
            "$ref": "#/components/schemas/Uzytkownik"
          },
          "ktoUsunalNavigation": {
            "$ref": "#/components/schemas/Uzytkownik"
          },
          "towars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Towar"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Klient": {
        "type": "object",
        "properties": {
          "idKlienta": {
            "type": "integer",
            "format": "int32"
          },
          "imie": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "nazwisko": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "idAdresu": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ktoDodal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kiedyDodano": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ktoZmodyfikowal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kiedyZmodyfikowano": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ktoUsunal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kiedyUsunieto": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "aktywny": {
            "type": "boolean",
            "nullable": true
          },
          "idAdresuNavigation": {
            "$ref": "#/components/schemas/Adre"
          },
          "ktoDodalNavigation": {
            "$ref": "#/components/schemas/Uzytkownik"
          },
          "ktoUsunalNavigation": {
            "$ref": "#/components/schemas/Uzytkownik"
          },
          "ktoZmodyfikowalNavigation": {
            "$ref": "#/components/schemas/Uzytkownik"
          }
        },
        "additionalProperties": false
      },
...
      "MetodaPlatnosci": {
        "type": "object",
        "properties": {
          "idMetodyPlatnosci": {
            "type": "integer",
            "format": "int32"
          },
          "nazwa": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "zamowienies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Zamowienie"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RolaUzytkownika": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SesjaKoszyka": {
        "type": "object",
        "properties": {
          "idSesjiKoszyka": {
            "type": "integer",
            "format": "int32"
          },
          "idKlienta": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "suma": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "dataUtworzenia": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "elementKoszykas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ElementKoszyka"
            },
            "nullable": true
          },
          "idKlientaNavigation": {
            "$ref": "#/components/schemas/Klient"
          }
        },
        "additionalProperties": false
      },
      "Towar": {
        "type": "object",
        "properties": {
          "idTowaru": {
            "type": "integer",
            "format": "int32"
          },
          "nazwa": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "kod": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "idKategorii": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cena": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "naStanie": {
            "type": "boolean",
            "nullable": true
          },
          "zdjecieUrl": {
            "type": "string",
            "nullable": true
          },
          "opis": {
            "type": "string",
            "nullable": true
          },
          "kiedyDodano": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ktoDodal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kiedyZmodyfikowano": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ktoZmodyfikowal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "kiedyUsunieto": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ktoUsunal": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "aktywny": {
            "type": "boolean",
            "nullable": true
          },
          "elementKoszykas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ElementKoszyka"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TowarZamowienium": {
        "type": "object",
        "properties": {
          "idTowaruZamowienia": {
            "type": "integer",
            "format": "int32"
          },
          "idTowaru": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idZamowienia": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ilosc": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "aktywny": {
            "type": "boolean",
            "nullable": true
          },
          "idTowaruNavigation": {
            "$ref": "#/components/schemas/Towar"
          },
          "idZamowieniaNavigation": {
            "$ref": "#/components/schemas/Zamowienie"
          }
        },
        "additionalProperties": false
      },
      "Uzytkownik": {
        "type": "object",
        "properties": {
          "idUzytkownika": {
            "type": "integer",
            "format": "int32"
          },
          "imie": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "nazwisko": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "dataDodania": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "email": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "zahaszowaneHaslo": {
            "type": "string",
            "nullable": true
          },
          "rolaUzytkownika": {
            "$ref": "#/components/schemas/RolaUzytkownika"
          },
          "rolaUzytkownikaId": {
            "type": "integer",
            "format": "int32"
          },
          "adreKtoDodalNavigations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Adre"
            },
            "nullable": true
          },
          "adreKtoEdytowalNavigations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Adre"
            },
            "nullable": true
          },
          "adreKtoUsunalNavigations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Adre"
            },
            "nullable": true
          },
          "kategoriumKtoDodalNavigations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Kategorium"
            },
            "nullable": true
          },
          "kategoriumKtoEdytowalNavigations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Kategorium"
            },
            "nullable": true
          },
          "kategoriumKtoUsunalNavigations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Kategorium"
            },
            "nullable": true
          },
          "klientKtoDodalNavigations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Klient"
            },
            "nullable": true
          },
          "klientKtoUsunalNavigations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Klient"
            },
            "nullable": true
          },
          "klientKtoZmodyfikowalNavigations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Klient"
            },
            "nullable": true
          },
          "towarKtoDodalNavigations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Towar"
            },
            "nullable": true
          },
          "towarKtoUsunalNavigations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Towar"
            },
            "nullable": true
          },
          "towarKtoZmodyfikowalNavigations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Towar"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Zamowienie": {
        "type": "object",
        "properties": {
          "idZamowienia": {
            "type": "integer",
            "format": "int32"
          },
          "dataZamowienia": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idKlienta": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "terminDostawy": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "suma": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "idMetodyPlatnosci": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idKlientaNavigation": {
            "$ref": "#/components/schemas/Klient"
          },
          "idMetodyPlatnosciNavigation": {
            "$ref": "#/components/schemas/MetodaPlatnosci"
          },
          "towarZamowienia": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TowarZamowienium"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}